Implemented item store to enable buying items when you have enough points. Buying items will also activate the interval abilities of increasing score by the defined amount.

+++ LemonClicker/Assets/GameManager.cs
@@ -9,7 +9,7 @@ public class GameManager : Singleton<GameManager>
     public ClickAction clickAction;
 
     /** The game points */
-    public int points;
+    public float points;
 
     /** The cooldown period in seconds regain limit */
     public int cooldown = 2;
